Deleting a Password needs an eTag since 2.5.0b1

+++ plugins/modules/password.py
@@ -189,7 +189,6 @@ class PasswordsUpdateAPI(CheckmkAPI):
 
 class PasswordsDeleteAPI(CheckmkAPI):
     def delete(self):
-
         return self._fetch(
             code_mapping=HTTP_CODES_DELETE,
             endpoint="/objects/password/%s" % self.params.get("name"),
@@ -270,10 +269,15 @@ def run_module():
             time.sleep(3)
 
     if module.params.get("state") == "absent":
-        passworddelete = PasswordsDeleteAPI(module)
-        result = passworddelete.delete()
+        passwordget = PasswordsGetAPI(module)
+        result = passwordget.get()
 
-        time.sleep(3)
+        if result.http_code == 200:
+            passworddelete = PasswordsDeleteAPI(module)
+            passworddelete.headers["If-Match"] = result.etag
+            result = passworddelete.delete()
+
+            time.sleep(3)
 
     module.exit_json(**result_as_dict(result))
